[refactor] move 'EmptyStateAnimation' into separate composable file
